Simplify jmx test plan, removes useless parts setup only one test plan line
there used to be 2almost similar lines because the initial plan was
testing the 2 stacks nodejs/java at the same time.
also initially the test was using a nignx in front of the servers, which
I don't know why, so I will remove it which makes me set all requests on
8080 instead of 4000
I also extract the auth token with a json extractor instead of regex.
stuff like that